% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseDeescalation-package.R
\docType{package}
\name{DoseDeescalation-package}
\alias{DoseDeescalation-package}
\alias{DoseDeescalation}
\title{The 'DoseDeescalation' package.}
\description{
This R package provides all of the code necessary to re-create
the simulation study reported in \emph{Targeted randomization de-escalation
trials enable fractional dosing of scarce drugs} by Boonstra, Tabarrok, and
Strohbehn.
}
\references{
Boonstra, Tabarrok, and Strohbehn (2023). Targeted randomization de-escalation
trials enable fractional dosing of scarce drugs. Under Review

Stan Development Team (2020). RStan: the R interface to Stan. R package
version 2.21.2. https://mc-stan.org
}
