% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_curves.R
\name{gen_curves}
\alias{gen_curves}
\title{Generate Hill equations}
\usage{
gen_curves(
  number_dose_levels = 6,
  a = 0.2,
  b_seq = c(0.85, 0.7, 0.55, 0.4),
  c = 0.7,
  d_seq = c(1, 2, 3, 4, 5, 7, 10, 20)
)
}
\arguments{
\item{number_dose_levels}{Positive integer giving number of doses per curve}

\item{a}{The 'a' parameter of the Hill equation. Must be a single number}

\item{b_seq}{The 'b' parameter. May be a vector, which will result in one curve per value}

\item{c}{The 'c' parameter. Must be a single number}

\item{d_seq}{The 'd' parameter. May be a vector, which will result in one curve per value}
}
\value{
A list with length \code{length(b_seq) * length(d_seq)}, each with \code{number_dose_levels}
elements, giving the true probabilities at equal-spaced points along the x-axis
}
\description{
Generate a list of curves from the Hill equations
}
