% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_dist_xi.R
\name{get_posterior_dist_xi}
\alias{get_posterior_dist_xi}
\title{Posterior sampling of \eqn{\xi}}
\usage{
get_posterior_dist_xi(params, n_draws = nrow(params)/2, data_grouped = NULL)
}
\arguments{
\item{params}{A tibble coming from calling \code{rstan::extract(x)$xi}, where
x is a stanfit object from either the horseshoe or horseshoe plus priors}

\item{n_draws}{An integer giving the desired number of draws from the posterior
distribution}

\item{data_grouped}{A tibble with three named columns: \code{n}, \code{resp}, and \code{dose_num},
corresponding to the number of observations, number of responses among those
observations, and dose number label}
}
\value{
A tibble with the same number of columns as \code{params}
}
\description{
Clever way to sample from the posterior of \eqn{\xi} given draws
from the prior and the log-likelihood of data
}
