% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_dist_xi.R
\name{get_posterior_dist_xi}
\alias{get_posterior_dist_xi}
\title{Posterior sampling of \eqn{\xi}}
\usage{
get_posterior_dist_xi(params, n_draws = nrow(params)/2, data_grouped = NULL)
}
\arguments{
\item{params}{A matrix that normally comes from calling
\code{rstan::extract(x)$xi}, where \code{x} is a stanfit object from either
\code{iso_horseshoe} or \code{iso_horseshoe_plus}. The rows of this matrix represent
draws from the \emph{prior} distribution, and the columns represent the
individual components of \eqn{\xi}.}

\item{n_draws}{An integer giving the desired number of draws from the
posterior distribution.}

\item{data_grouped}{A tibble with two named columns: \code{n} and \code{resp},
corresponding to the number of observations and number of responses among
those observations. The number of rows must be equal to the number of
columns in \code{params}, corresponding to the number of dose levels. It is
assumed that the rows are ordered by dose number, with the first row
corresponding to the smallest/lowest dose and the last row corresponding to
the largest/highest dose. Additional columns are allowed and will be
ignored. \code{data_grouped} can alternatively be \code{NULL}, in which case you will
get draws from the priors.}
}
\value{
A matrix containing draws from the posterior. It will have the same
number of columns as \code{params} and \code{n_draws} rows.
}
\description{
Clever way to sample from the posterior of \eqn{\xi} given draws
from the prior and the log-likelihood of data
}
