% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_prob_mdse.R
\name{get_posterior_prob_mdse}
\alias{get_posterior_prob_mdse}
\title{Posterior probability of being minimum dose with satisfactory efficacy (MDSE)}
\usage{
get_posterior_prob_mdse(params, relative_threshold, data_grouped = NULL)
}
\arguments{
\item{params}{A tibble coming from calling \code{rstan::extract(x)$xi}, where
x is a stanfit object from either the horseshoe or horseshoe plus priors.}

\item{relative_threshold}{A number between 0 and 1 indicating the smallest
relative efficacy as compared to the largest dose level that would still
be considered acceptable.}

\item{data_grouped}{A tibble with three named columns: \code{n}, \code{resp}, and \code{dose_num},
corresponding to the number of observations, number of responses among those
observations, and dose number label}
}
\value{
A simplex vector of probabilities
}
\description{
Clever way to calculate the posterior probability of
each dose level being the MDSE given draws
from the prior and the log-likelihood of data
}
